rag-system/pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "RAG system for intelligent product search and recommendations"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.20",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.2",
    "openai>=1.10.0",
    "tiktoken>=0.5.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
